deixando a visualização  bonita e padronizando o codigo

+++ metodos.py
@@ -1,5 +1,6 @@
 from operacoesbd import *
 
+# função para verificar se um número é válido
 def validarNumero(num):
     try:
         num = int(num)
@@ -7,47 +8,43 @@ def validarNumero(num):
     except ValueError:
         return False
 
-
-
-
+# função que solicita a escolha de uma categoria e a retorna
 def escolherCategoria():
-        print("======================")
-        print("      CATEGORIAS      ")
-        print("======================")
-        print("1) Reclamação\n2) Sugestão\n3) Elogios")
+        print(f"\n{"="*22}\n      CATEGORIAS      \n{"="*22}\n ")
+        print("1) Reclamação\n2) Sugestão\n3) Elogios\n")
 
         escolha = input("Escolha a categoria da manifestação: ")
         
-        if validarNumero(escolha):
-            escolha = int(escolha)
-        else: 
-            print("⚠️  Digite uma opção válida!")
-
-        if escolha > 0 and escolha < 4:
-            if escolha == 1:
-                categoria = "reclamação"
-            elif escolha == 2:
-                categoria = "sugestão"
-            elif escolha == 3:
-                categoria = "elogio"
-
-            return categoria
-        else:
-            print("\n⚠️  Digite uma opção válida!")
+        if not validarNumero(escolha):
+            print("\n⚠️  Digite uma opção válida!\n")
             return ""
+        else: 
+            escolha = int(escolha)
+
+            if escolha > 0 and escolha < 4:
+                if escolha == 1:
+                    categoria = "reclamação"
+                elif escolha == 2:
+                    categoria = "sugestão"
+                elif escolha == 3:
+                    categoria = "elogio"
+
+                return categoria
+            else:
+                print("\n⚠️  Digite uma opção válida!\n")
+                return ""
+
+# função para listar todas as manifestações
+def listarManifestacoes(conn):
+    sql = "select * from manifestacoes"
+
+    # pega todas as manifestações cadastradas no bd
+    manifestacoes = listarBancoDados(conn, sql)
+
+    return manifestacoes
 
 
-
-
-def criarManifestacao(conn, categoria, assunto):
-    
-    sql = "insert into manifestacoes(categoria, assunto)values(%s,%s)"
-    dados = [categoria, assunto]
-
-    insertNoBancoDados(conn, sql, dados)
-
-
-
+# função para listar todas as manifestações da categoria escolhida 
 def listarManifestacoesCategoria(conn, categoria):
     sql = "select * from manifestacoes where categoria = %s"
     dados = [categoria]
@@ -57,55 +54,67 @@ def listarManifestacoesCategoria(conn, categoria):
     if len(manifestacoes) == 0:
         print("\n❌ Não há manifestações cadastradas!")
     else:
-        print("\nLista de manifestações:")
         for manifestacao in manifestacoes:
-            print("\n💠 Categoria:", manifestacao[1], "\nCódigo:", manifestacao[0],"\nAssunto:",manifestacao[2])
+            print(F"▫️  Categoria: {manifestacao[1]}\nCódigo: {manifestacao[0]}\nAssunto: {manifestacao[2]}\n")
 
 
+# função para criar uma nova manifestação
+def criarManifestacao(conn, categoria, assunto):
+    
+    sql = "insert into manifestacoes(categoria, assunto)values(%s,%s)"
+    dados = [categoria, assunto]
 
-def pesquisarCodigo (conn):
-    while True:
-        codigo = input("\nInforme o codigo para pesquisa: ")
-        if not codigo.isdigit():
-            print("\n❌ Digmanifestacao novamente em numeral")
-            continue
-        break
-    codigo = int(codigo)
-    perquisarcodigo = "select * from manifestacoes where id = (%s)"
-    dados = [codigo]
-    manifestacao = listarBancoDados(conn, perquisarcodigo, dados)
-
-    if len(manifestacao) == 0:
-        print("\n⚠️ Não contem item na lista")
-    else:
-        print("\n🔎 codigo:", manifestacao[0][0], "\nTipo:", manifestacao[0][1],"\nManifestação:", manifestacao[0][2] )
+    insertNoBancoDados(conn, sql, dados)
 
 
-
-
-def remover(conn):
-    codigoRemocao = int(input("\n🗑 Qual o codigo a ser removido? "))
-    consultaRemocao = "delete from manifestacoes where id = %s"
-    dadosRemocao = [codigoRemocao]
-
-    excluirBancoDados(conn,consultaRemocao,dadosRemocao)
-    print("\n✅ Removido com sucesso!")
-
-
-
-
-def listarManifestacoes(conn):
-    sql = "select * from manifestacoes"
-
-    manifestacoes = listarBancoDados(conn, sql)
-
-    return manifestacoes
-
-#Exibir quantidade de informações
+# função para exibir a quantidade de manifestações cadastradas no bd
 def exibirQuantidadeManifestacoes(conn):
     sql = "select count(*) from manifestacoes"
 
+    # pega o número de manifestação cadastradas no bd
     quantidadeManifestacoes = listarBancoDados(conn, sql)
 
     return quantidadeManifestacoes
 
+
+# função para pesquisar um manifestação pelo código
+def pesquisarManifestacaoCodigo (conn):
+    codigoPesquisa = input("Informe o codigo para pesquisa: ")
+    
+      # verifica se o código digitado é um número válido
+    if not validarNumero(codigoPesquisa):
+        print("\n⚠️  Digite um código válido!\n")
+    else:
+        # se o código digitado for um número válido ele executa as linhas debaixo
+        codigoPesquisa = int(codigoPesquisa)
+        sql = "select * from manifestacoes where id = (%s)"
+        dados = [codigoPesquisa]
+        manifestacao = listarBancoDados(conn, sql, dados)
+
+        if len(manifestacao) == 0:
+            print("\n❌ Manifestação não encontrada!\n")
+        else:
+            print(f"\n🔎 Categoria: {manifestacao[0][1]}\nAssunto: {manifestacao[0][2]}\nCódigo: {manifestacao[0][0]}\n")
+
+
+# função para remover uma manifestação
+def removerManifestacao(conn):
+    codigoRemocao = input("Digite o código da manifestção a ser removida: ")
+    
+    # verifica se o código digitado é um número válido
+    if not validarNumero(codigoRemocao):
+        print("\n⚠️  Digite um código válido!\n")
+    else:
+        # se o código digitado for um número válido ele executa as linhas debaixo
+        codigoRemocao = int(codigoRemocao)
+
+        sql = "delete from manifestacoes where id = %s"
+        dados = [codigoRemocao]
+
+        linhasAfetadas = excluirBancoDados(conn,sql,dados)
+        
+        # verifica se houve alguma alteração no bd
+        if linhasAfetadas == 0:
+            print("\n❌ Nenhuma manifestação removida!\n")
+        else:
+            print("\n✅ Manifestação removida com sucesso!\n")
